# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PartSignal API
  version: 0.1.1
  description: PartSignal MVP 契约。除登录外，写请求必须携带 X-CSRF-Token。
servers:
  - url: /
security:
  - sessionCookie: []
...
paths:
...
  /api/v1/files/upload-intents:
    post:
      tags: [files]
      operationId: createFileUploadIntent
      parameters: [{$ref: '#/components/parameters/CsrfHeader'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UploadIntentCreate'}
      responses:
        '201':
          description: 已创建限时上传意图
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UploadIntent'}
  /api/v1/files/{file_id}/complete:
    post:
      tags: [files]
      operationId: completeFileUpload
      parameters:
        - {$ref: '#/components/parameters/FileId'}
        - {$ref: '#/components/parameters/CsrfHeader'}
      responses:
        '200':
          description: HEAD 校验成功
          content:
            application/json:
              schema: {$ref: '#/components/schemas/FileRecord'}
        '422': {$ref: '#/components/responses/ErrorResponse'}
  /api/v1/files/{file_id}:
    get:
      tags: [files]
      operationId: getFileRecord
      parameters: [{$ref: '#/components/parameters/FileId'}]
      responses:
        '200':
          description: 文件元数据
          content:
            application/json:
              schema: {$ref: '#/components/schemas/FileRecord'}
  /api/v1/files/{file_id}/abort:
    post:
      tags: [files]
      operationId: abortFileUpload
      parameters:
        - {$ref: '#/components/parameters/FileId'}
        - {$ref: '#/components/parameters/CsrfHeader'}
      responses:
        '200':
          description: 已中止未完成上传
          content:
            application/json:
              schema: {$ref: '#/components/schemas/FileRecord'}
        '409': {$ref: '#/components/responses/ErrorResponse'}
  /api/v1/files/{file_id}/download-url:
    get:
      tags: [files]
      operationId: getFileDownloadUrl
      parameters: [{$ref: '#/components/parameters/FileId'}]
      responses:
        '200':
          description: 短期下载 URL
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SignedUrl'}
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: partsignal_session
  parameters:
    CsrfHeader:
      name: X-CSRF-Token
      in: header
      required: true
      schema: {type: string, minLength: 32}
...
    FileId: {name: file_id, in: path, required: true, schema: {type: string, format: uuid}}
...
  responses:
    ErrorResponse:
      description: 业务或校验错误
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
  schemas:
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error: {$ref: '#/components/schemas/ErrorDetail'}
    ErrorDetail:
      type: object
      additionalProperties: false
      required: [code, message, request_id]
      properties:
        code: {type: string}
        message: {type: string}
        details: {type: object, additionalProperties: true, default: {}}
        request_id: {type: string}
...
    Confidentiality: {type: string, enum: [PUBLIC, INTERNAL, RESTRICTED]}
...
    FileCategory: {type: string, enum: [EVIDENCE, OPERATION_SCREENSHOT, PUBLICATION_ASSET]}
    UploadIntentCreate:
      type: object
      additionalProperties: false
      required: [category, original_filename, content_type, size, sha256, access_level]
      properties:
        category: {$ref: '#/components/schemas/FileCategory'}
        original_filename: {type: string}
        content_type: {type: string}
        size: {type: integer, minimum: 1}
        sha256: {type: string, pattern: '^[a-f0-9]{64}$'}
        access_level: {$ref: '#/components/schemas/Confidentiality'}
    UploadInstruction:
      type: object
      additionalProperties: false
      required: [method, url, headers, fields, expires_at]
      properties:
        method: {type: string, enum: [PUT, POST]}
        url: {type: string, format: uri, minLength: 1, maxLength: 2083}
        headers: {type: object, additionalProperties: {type: string}}
        fields: {type: object, additionalProperties: {type: string}}
        expires_at: {type: string, format: date-time}
    UploadIntent:
      type: object
      additionalProperties: false
      required: [file, upload]
      properties:
        file: {$ref: '#/components/schemas/FileRecord'}
        upload: {$ref: '#/components/schemas/UploadInstruction'}
    FileRecord:
      type: object
      additionalProperties: false
      required: [id, category, original_filename, object_key, content_type, size, sha256, access_level, status, created_at]
      properties:
        id: {type: string, format: uuid}
        category: {$ref: '#/components/schemas/FileCategory'}
        original_filename: {type: string}
        object_key: {type: string}
        content_type: {type: string}
        size: {type: integer}
        sha256: {type: string}
        access_level: {$ref: '#/components/schemas/Confidentiality'}
        status: {type: string, enum: [PENDING, VERIFIED, FAILED, ABORTED]}
        created_at: {type: string, format: date-time}
        verified_at: {type: [string, 'null'], format: date-time}
    SignedUrl:
      type: object
      additionalProperties: false
      required: [url, expires_at]
      properties:
        url: {type: string, format: uri, minLength: 1, maxLength: 2083}
        expires_at: {type: string, format: date-time}